+++ test/users.js
@@ -1,6 +1,6 @@
 const expect = require("chai").expect;
 const usersHandlers = require("../src/services/users");
-const { userModel } = require("../src/db/mongo");
+const { userModel } = require("../src/modules/stores/mongo");
 const { userGenerator, testRemover } = require("./utils");
 const httpMocks = require("node-mocks-http");
